backend/requirements.txt
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.1
certifi==2026.2.25
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
jiter==0.13.0
openai==2.28.0
packaging==26.0
pillow==11.3.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
sniffio==1.3.1
tomli==2.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
pymupdf>=1.26.0
jinja2>=3.1
xhtml2pdf>=0.2.11
